refactored reads.js organization
