+++ mango-cli/src/main/webapp/WEB-INF/layouts/reads.ssp
@@ -1,4 +1,5 @@
 <%@ val viewRegion: (String, String, String) %>
+<%@ val samples: (String, String) %>
 
 <html>
 <head lang="en">
@@ -7,6 +8,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1">
     <meta name="description" content="Reads Visualization for Mango">
     <meta name="author" content="Eric Tu">
+    <script src="https://ajax.googleapis.com/ajax/libs/jquery/1.11.3/jquery.min.js"></script>
     <script src="http://d3js.org/d3.v3.min.js" charset="utf-8"></script>
     <link rel="stylesheet" href="//maxcdn.bootstrapcdn.com/bootstrap/3.3.4/css/bootstrap.min.css">
     <link href='http://fonts.googleapis.com/css?family=Open+Sans:300,400,600,700' rel='stylesheet' type='text/css'>
@@ -120,17 +122,19 @@
 
 <br>
 
-<div id="refArea"></div>
-<div id="readsArea"></div>
+<div class="graphArea" id="refArea"></div>
+<div class="graphArea" id="readsArea"></div>
 <br>
 
 <script>
   var viewRefName = "${viewRegion._1}";
   var viewRegStart = parseInt(${viewRegion._2}, 10);
   var viewRegEnd = parseInt(${viewRegion._3}, 10);
+  var samp1Name = "${samples._1}";
+  var samp2Name = "${samples._2}";
   var readsExist = true
   var variantsExist = false
   var featuresExist = false
 </script>
-<script src="js/overall.js"></script>
+<script src="js/reads.js"></script>
 </html>
