+++ multsample.sh
@@ -1,2 +1,1 @@
-bin/mango-submit ./workfiles/mm10_chrM.fa chrM -sample1 person1 -read_file1 ./workfiles/mouse_chrM.bam -sample2 person2 -read_file2 ./workfiles/mouse_chrM.bam  -var_file ./workfiles/truetest.vcf.adam -feat_file ./workfiles/smalltest.bed
-
+bin/mango-submit ./workfiles/mm10_chrM.fa chrM -sample1 person1 -read_file1 ./workfiles/mouse_chrM.bam.adam -sample2 person2 -read_file2 ./workfiles/mouse_chrM.bam.adam  -var_file ./workfiles/truetest.vcf.adam -feat_file ./workfiles/smalltest.bed
